+++ .eslintrc.json
@@ -38,6 +38,7 @@
         "buddy_list": false,
         "channel": false,
         "click_handlers": false,
+        "color_data": false,
         "colorspace": false,
         "common": false,
         "components": false,
