+++ frontend_tests/node_tests/color_data.js
@@ -1,6 +1,4 @@
-global.stub_out_jquery();
-
-zrequire('stream_color');
+zrequire('color_data');
 
 run_test('pick_color', () => {
     var used_colors = ["#76ce90", "#fae589"];
@@ -8,7 +6,7 @@ run_test('pick_color', () => {
     // Colors are assigned randomly, so this test is a little vague and brittle,
     // but we can verify a few things, like not reusing colors and making sure
     // the color has length 7.
-    var color = stream_color.pick_color(used_colors);
+    var color = color_data.pick_color(used_colors);
     assert.notEqual(color, "#76ce90");
     assert.notEqual(color, "#fae589");
     assert.equal(color.length, 7);
