+++ static/js/archive.js
@@ -23,7 +23,7 @@ exports.initialize = function () {
     var recipient_and_topic = $('#display_recipient').html();
     var stream_name = recipient_and_topic.split('-')[0];
     var topic = recipient_and_topic.split('-')[1];
-    var recipient_color = stream_color.pick_color([]);
+    var recipient_color = color_data.pick_color([]);
     current_message_group.message_containers = [];
     current_message_group.show_date_separator = false;
     current_message_group.display_recipient = stream_name;
